plugins: Updated version of Kapost Byline plugin v1.0.5-WIP
* added protected _kapost_analytics custom field 
* added the_content filter 
* added the_content_feed filter 
* the filters append the kapost specific analytics 'code' to the 'content' based on the _kapost_analytics custom field

To summarize, our customers use us for post specific analytics. Generally that is javascript that is added at the foot of the post content. But using the WYSIWIG editor, it's possible to accidentally remove the analytics code. So we've added a safety check to re-add the analytics code if it is missing. This only applies to posts that have the _kapost_analytics custom field, which came from Kapost with the intention of tracking.

/z 2179


git-svn-id: https://vip-svn.wordpress.com/plugins@57010 dd790b09-5c1c-0410-8744-875116bcd8a6