WP.com Thumbnail Editor: Disable the functionality for non-cropping thumbnails for the time being.
Many themes register thumbnails with the dimensions of something like 300 x 99999 so that they get a 300 pixel wide image regardless of the image's height (no cropping). This confuses the plugin because it thinks you want a 300x99999 pixel thumbnail image. So for now, ignore non-cropping thumbnails. They get resized proportionally anyway so that the whole image can be seen.

At some point I'd like to allow controlling non-cropping thumbnails so that you can use a portion of an image as it's thumbnail (to account for the smaller image size) but for now it just causes too many problems.


git-svn-id: https://vip-svn.wordpress.com/plugins@68709 dd790b09-5c1c-0410-8744-875116bcd8a6