Photon doesn't support query strings so we strip them when generating the Photon URL.
However some images are actually being served by a PHP script using query string variables, such as phpThumb.php, so to help make sure an image shows up, check to make sure we aren't passing a .php script to Photon.

The best way to do this would be to make sure the extension matches Photon's whitelist but what if Photon's list changes in the future? Don't want to have to maintain two copies of that list.


git-svn-id: https://vip-svn.wordpress.com/plugins@78296 dd790b09-5c1c-0410-8744-875116bcd8a6