Co-Authors Plus: Add 'coauthors_post_author_value' as a filter for changing the 'post_author' value for a post.
This, for instance, allows you to set the 'post_author' to a default value you're using guest authors on a post.

The deficiency of this filter is that 'coauthors_update_post' fires later, so if you use get_coauthors() you might not be getting the most r
ecent values. Buyer beware.


git-svn-id: https://vip-svn.wordpress.com/plugins@78369 dd790b09-5c1c-0410-8744-875116bcd8a6