add-meta-tags:
Add a filter for the fields which will allow for additional fields to be injected or the fields to be reordered. 
Normalize CSS names (form_field and form-field) and add additional markup making it easier to customize the contents of the meta box.

Rationale: We want to have all fields included, but 99% of the time we're only using the Title field. With consistent CSS and a little extra markup we can (easily) accomplish what you see in the screenshot.

props coreygilmore


git-svn-id: https://vip-svn.wordpress.com/plugins@87576 dd790b09-5c1c-0410-8744-875116bcd8a6