VIP: Disqus: don't load the Disqus JS unless a post is published
The Disqus JS makes a connection to disqus.com when loaded, and uses the current URL and other information to create a new comment thread record
in their database. Query params are stripped from the URL, so previewed posts cause a new comment thread to be created, but all of those threads
are linked back to the root of the site. This can cause problems with accurately retrieving comments from Disqus and with other parts of the Disqus API.



git-svn-id: https://vip-svn.wordpress.com/plugins@101940 dd790b09-5c1c-0410-8744-875116bcd8a6