Bitly: Generate Bitly urls earlier for plugins that expect them regardless of post status
Previously, Bitly urls were generated on publish_post, so were not available before then. This caused problems with plugins that relied on the short url early, such as Publicize, which generates its message on save regardless of status. 

If a post was set to private then later published, the url saved in Publicize would be wp.me, but on publish a Bitly url is generated, resulting in Tweets containing both links (due to the Twitter Publicize service's implementation)


git-svn-id: https://vip-svn.wordpress.com/plugins@103194 dd790b09-5c1c-0410-8744-875116bcd8a6