Plugins: Edit Flow: Fix bug that was preventing permalink slugs from being edited
Previously, the fix_editable_slug() hook was called for both the sample permalink (slug editor) box and the postmeta box, the latter causing $post->post_name to be set. This resulted in post slugs not auto updating as the title changed, so we merged a GitHub Pull Request to get around it.

The PR introduced a new bug that prevented the slug from being editable at all (would always go back to the title). This correctly removes the filter in the edit_form_after_title hook. Previously, fix_editable_slug() attempted to do this, but if no actions were registered to the hook, the code to remove it was never fired leading to the original behavior problems. This reverts the PR and solves the original problem in the correct way.


git-svn-id: https://vip-svn.wordpress.com/plugins@107125 dd790b09-5c1c-0410-8744-875116bcd8a6