Plugins: Thumbnail Editor: Improve get_coordinates() to return coordinates if size is in a ratio group but coordinates haven't been explicitly saved for that size
Since this plugin saves the coordinates for every size in a ratio group in meta, adding a new size presents a problem for existing content - the size is not backfilled with coordinates.

Fortunately, the coordinates are very ratio specific, and not specific to the size. This means we can use the coordinates of any other size in the ratio group for unknown sizes, and they will crop correctly.


git-svn-id: https://vip-svn.wordpress.com/plugins@110908 dd790b09-5c1c-0410-8744-875116bcd8a6