Update to reflect github updates through 94ebec8a8f1f547e9e98d5c42f73f60ad719b157.
Adds multifield support, support for a couple of new field types, miscellaneous tweaks to readme, and inits on admin_init rather than init.

Note: If a theme's custom fields are initiated using the wrong hook (that is, not custom_metadata_manager_init_metadata), they won't display properly in the admin.



git-svn-id: https://vip-svn.wordpress.com/plugins@111312 dd790b09-5c1c-0410-8744-875116bcd8a6