Plugins: Co Authors Plus: Add new method for retreiving avatars for Co Authors and Guest Authors
Also refactors the custom Guest Authors table to use the new method, instead of get_avatar(), which doesn't return the correct avatar when 2 or more Guest Authors share an email, but have different thumbnails attached.

This overcomes the above limitation in the get_avatar() filter, which only accepts email addresses.


git-svn-id: https://vip-svn.wordpress.com/plugins@121506 dd790b09-5c1c-0410-8744-875116bcd8a6