Fixed regex: parsing filename from image source breaks when there's a query present, %name placeholder doesn't get replaced in this case.
The regex was looking only for the extension then end of the line. If the extension is followed by a '?', the regex fails to match. We need to look for '?', as well as the end of the line here.

Example of image source that wouldn't match prior to fix: http://blogduluthtrading.files.wordpress.com/2013/09/duluth-women-model-nora-in-waxed-cotton-jacket.jpg?w=213&h=213&crop=1


git-svn-id: https://vip-svn.wordpress.com/plugins@124164 dd790b09-5c1c-0410-8744-875116bcd8a6