Fix notice in get_page_by_title call.
We pass the same $object_type that the developer wants but expect an array.

Since we don't end up using the fetched object, we can just request an OBJECT.

It would be more efficient to not fetch two versions of the page on an uncached call, but trying to make it work for all the different data types will be tricky.


git-svn-id: https://vip-svn.wordpress.com/plugins@124339 dd790b09-5c1c-0410-8744-875116bcd8a6