Adbusters plugin: add support for non-root paths
To help the transition to Adbusters, a filter has been added to let you serve the files from non-root URLs.
The plugin now also checks that the template exists before trying to load it.

If you've used the wpcom_vip_maybe_load_ad_busters filter previous to this commit, note that the arguments passed to this function have changed slightly; check the source.


git-svn-id: https://vip-svn.wordpress.com/plugins@126148 dd790b09-5c1c-0410-8744-875116bcd8a6