Remove conditional that adds the current post id to the stats query
This function should only rely on it's args, not globals, to produce a result. These results are cached elsewhere, so depending on when it is called, the cache can get corrupted with results from a single post.

Also, there is a function, wpcom_vip_get_post_pageviews() for stats by post. This shouldn't be silently modifying the query - operate on the args, and only the args.


git-svn-id: https://vip-svn.wordpress.com/plugins@130105 dd790b09-5c1c-0410-8744-875116bcd8a6