Adds support for higher requests per hour by making the hourly limit filterable and batching the query.
The query now processes posts in batches of 100, up to the value specified by the 'bitly_hourly_limit' filter (defaults to 100). Useful for large sites that need to process at a faster rate.

Also adds a utility bitly_log() function (for use in CLI scripts) and refactors the 'mark as complete' code into a separate function.

props @bradfordcampeaulaurion


git-svn-id: https://vip-svn.wordpress.com/plugins@130760 dd790b09-5c1c-0410-8744-875116bcd8a6