wpcom-legacy-redirector: allow redirecting to non-post URLs (like categories and terms).
Allow the insert method to accept a non-post redirect_to value and store that in the post_excerpt field.

Also updates docs to note these new features and updates cache group to flush caches since we now store the ID of the redirect post, not the post parent.

h/t moguldomrich (for the initial idea and earlier code)


git-svn-id: https://vip-svn.wordpress.com/plugins@133769 dd790b09-5c1c-0410-8744-875116bcd8a6