Elasticsearch wpcom_search_api_wp_to_es_args(): add cross_fields to standard multi_match query.
cross_fields option is new in ES 1.x and allows terms in an AND query to match across different fields.

For instance searching "wordpress jetpack" with an AND multi_match on title and content fields can now match posts where jetpack is in the title and wordpress is in the content. Previously there was no easy way to make this work.

(This change has been live in the wp.com version of this code for two weeks)



git-svn-id: https://vip-svn.wordpress.com/plugins@153178 dd790b09-5c1c-0410-8744-875116bcd8a6