Edit Flow: fix incorrect handling where EF subscribers, whose username couldn't be matched to a wp.com account or who no longer had access to the sender blog, weren't being removed from notification email's "to" list.
The `get_following_users` method is re-using the same variable to fetch a list of subscribers' user names, and then validate whether an email should be sent to each user.
The checks around the current `is_user_member_of_blog` do stop emails being sent to users who no longer had access to the blog, but that user record wasn't being unset from the return array.

This resulted in many failed emails because the "To" address was someone's wp.com username.

See https://99deploys.wordpress.com/2014/04/22/there-are-a-couple-problems/ for some background.


git-svn-id: https://vip-svn.wordpress.com/plugins@165810 dd790b09-5c1c-0410-8744-875116bcd8a6