Fixing SSL related issue in API calls
This commit replaces hardcoded http endpoint urls with protocol relative ones in order to make the plugin working properly in SSL powered administration.
A minor logic tweak was needed in order to make the plugin working correctly in cases when a session has to be created/used. That was needed as there is no longer "http" in the url.



git-svn-id: https://vip-svn.wordpress.com/plugins@182707 dd790b09-5c1c-0410-8744-875116bcd8a6