Preventing import status checking all the comments on the site from being triggered outside the plugin's page
detect_default_comment function can be pretty resource demanding the import status is only available to user when on the plugin's page. So there is no point on checking the process on every - even frontend - page load since the begin() function is also being triggered on that page only.

I also swapped the conditions in the if clause in order to not trigger the detect_default_comment function unnecessarily since get_opton is much much faster.


git-svn-id: https://vip-svn.wordpress.com/plugins@193871 dd790b09-5c1c-0410-8744-875116bcd8a6