Fixing "Missing argument 2 for wp_flush_term_exists()" warning and making the function actually work as expected
In order to be able to properly flush cache introduced by `wpcom_vip_term_exists` function on terms deletion I hooked the function to a later action ( `delete_term` vs. original `deleted_term_taxonomy` ) as that one provides us with a copy of original term object and we can get all needed fields from that place.

I also fixed the `id` in foreach array which has to state `term_id` as `id` is not a property of term object.

Fixes #39346-z


git-svn-id: https://vip-svn.wordpress.com/plugins@195803 dd790b09-5c1c-0410-8744-875116bcd8a6