VIP Plugins: Added basic ability to load RC plugins when using the VIP Plugins UI.
Add the blog sticker "vip-plugins-ui-rc-plugins" to a site and any plugins loaded via the UI that have a release candidate will load that version instead.

Due to the potential unstable nature of RCs, this has a chance of causing data loss.  Be careful and/or only use this on test sites when possible.


git-svn-id: https://vip-svn.wordpress.com/plugins@219908 dd790b09-5c1c-0410-8744-875116bcd8a6