VIP Plugins: Added basic ability to load RC plugins when using wpcom_vip_load_plugin()
Add the blog sticker "vip-plugins-ui-rc-plugins" to a site and any plugins loaded via code that have a release candidate will load that version instead.

Due to the potential unstable nature of RCs, this has a chance of causing data loss. Be careful and/or only use this on test sites when possible.

Also see r219908 for loading RC plugins in via the Plugins UI


git-svn-id: https://vip-svn.wordpress.com/plugins@219945 dd790b09-5c1c-0410-8744-875116bcd8a6