Delete UppSite plugin; the service has been shut down and the domain no longer points to the correct place.
One site is listed as still using the plugin, but on investigating that site it doesn't appear to be doing so at all (elev8.hellobeautiful.com). I can find no clients who are activating the plugin in code.

We should remove it before a domain hijacker tries to do something with the plugin panel in case there is a client who has it installed in some way that the MC page is not reading.


git-svn-id: https://vip-svn.wordpress.com/plugins@221774 dd790b09-5c1c-0410-8744-875116bcd8a6