WP SEO: adding check for catching WP_Error objects passed to WP_SEO::meta_field method
WP_SEO::format method can return a WP_Error object in case it's not provided with string attribute so we have to check for WP_Error in WP_SEO::meta_field in order to prevent Catchable fatal errors.

This fix is using simple is_string check. If it fails, the variable is turned into an empty string.


git-svn-id: https://vip-svn.wordpress.com/plugins@226754 dd790b09-5c1c-0410-8744-875116bcd8a6