Babble RC: fixing handling page translation for permalink in a form of %postname%
While we can't filter the initial parse_request we need to take some guesses in order to figure out a page is actually being requested.

The code in this revision is taking advantage of core's code which can be found in WP::parse_request

In case of slug collision of post and page, the page has got a precedence - that's the same how core handles this situation.


git-svn-id: https://vip-svn.wordpress.com/plugins@232591 dd790b09-5c1c-0410-8744-875116bcd8a6