vip-helper-stast-wpcom: fixing sort bug in wpcom_vip_stats_csv_print
There is a bug in case there are two or more posts with same number of postviews. In such case, only one post of those with the same number of postviews is added to the final array.

This changeset fixes that bug for all functions taking advantage of wpcom_vip_stats_csv_print function. Namely:

* wpcom_vip_top_posts_array
* wpcom_vip_get_stats_array (if $table argument is postviews)
* wpcom_vip_get_stats_csv (if $table argument is postviews)

P2: https://keepingtheirblogsgoing.wordpress.com/2016/02/29/fixes-coming-to-wpcom_vip_top_posts_array-today/


git-svn-id: https://vip-svn.wordpress.com/plugins@234680 dd790b09-5c1c-0410-8744-875116bcd8a6