Most Popular Feed: move the `get_option( "rewrite_rules" )` to inside the conditional in order to make the plugin to work properly in WP CLI and WP Rest API enabled on the WordPress.com platform
As we're filtering the rewrite_rules option through the WordPress.com compatibility plugin for the WP Rest API and we're currently not loading the WP Rest API functions in WP CLI, the option_rewrite_rules filter's callback is throwing Fatal Errors if this plugin is enabled.

Moving the get_option to inside the conditional solves the issue.


git-svn-id: https://vip-svn.wordpress.com/plugins@245479 dd790b09-5c1c-0410-8744-875116bcd8a6