VIP Search Add-On : Order of date_histogram facets
Summary: Clients have requested the ability to show date_histogram facets from newest to oldest.  This patch enables an 'order' property of the date_histogram facet to implement ordering. The property gets ignored when creating the ES Query but then gets used when returning data to make sure the order is applied.

Differential Revision: https://code.a8c.com/D5092


git-svn-id: https://vip-svn.wordpress.com/plugins@290470 dd790b09-5c1c-0410-8744-875116bcd8a6