VIP: Prevent double-loading of plugins.
Summary:
It's possible to load a plugin twice (or more!) on VIP. This patch builds on patches by @batmoo and @davidbinovec to prevent plugins loading twice.

We use a static variable to record all activated plugins each time `wpcom_vip_load_plugin()` is called. This list is checked on every call to ensure the plugin hasn't already been loaded. If it has we alert the VIP team via IRC so we can proactively help VIPs and we also warn admins on the relevant sites using an admin notice.

Test Plan:
Apply the patch in a sandbox and;

* Activate a plugin via the UI and code
* Activate a plugin via the UI and a different version by code
* Activate a plugin via code twice
* Activate a plugin via code and then another version of the same plugin by code
* Do all of the above at the same time. ALL THE PLUGINS!
* Activate a plugin once via UI
* Activate a plugin once via code

Check that in all but the last two cases the IRC message is sent, an admin notice appears and the plugin does not load more than once.

Check that in the last two cases the plugin loads successfully.

Reviewers: davidbinovec, batmoo

Reviewed By: davidbinovec

Subscribers: batmoo, davidbinovec, #vip_developer_squad

Differential Revision: https://code.a8c.com/D5367



git-svn-id: https://vip-svn.wordpress.com/plugins@295135 dd790b09-5c1c-0410-8744-875116bcd8a6