es-wp-query (RC): Allow adapters to normalize term value (like `strtolower` if mapping to `raw_lc`).
As some adapters, like wpcom-vip, are using normalised value for mapping term name, we need to allow them to normalise the value passed to Elasticsearch. This commit is adding a filter which allows modifications of the value and is adding a callback which takes advantage of the new filter for fixing an issue in the wpcom-vip adapter.


git-svn-id: https://vip-svn.wordpress.com/plugins@296787 dd790b09-5c1c-0410-8744-875116bcd8a6