WP REST API: Add workaround for CAP author detection breaking embed requests.
In REST API context, the `count_user_posts` function should work as originally intended in core, since it is linked to the actual `post_author` property of the post, rather than the author abstraction that CAP adds.

In the future, this should be addressed within CAP by (1) disabling this functionality during REST API requests and (2) providing full-featured embeds for multiple authors.


git-svn-id: https://vip-svn.wordpress.com/plugins@301705 dd790b09-5c1c-0410-8744-875116bcd8a6