VIP: Remove adding of option if not present.
When connection problems happen with memcache, some get_option calls might fail. This leads to WordPress's add_option to kick in and overwrite the option in the database even if it's already present. 
Instead we'll rely on the default value being used (L558) of a blank array which while it would not load the plugins for that specific request will atleast not overwrite it with blank data.  


git-svn-id: https://vip-svn.wordpress.com/plugins@301908 dd790b09-5c1c-0410-8744-875116bcd8a6