VIP: wpcom_vip_file_get_content bug fix
Currently if the request fails AND we have content in the backup cache key the code will not set the flag that the remote server is down. This means the next run of the function will also query the remote host. This patch changes that to add a 60s delay to the next remote query regardless of if the backup data is found or not. It also removes the legacy keys and changes the format of the backup and disable key, this change is to invalidate the caches that could cause use to keep hitting the remote site over and over again. (It also fixed a problem during testing where the keys were never found, it was initially thought this was related to key length but this could not be replicated outside of this function)

Differential Revision: https://code.a8c.com/D6805


git-svn-id: https://vip-svn.wordpress.com/plugins@306713 dd790b09-5c1c-0410-8744-875116bcd8a6