VIP: Prevent double-loading of plugins.
It's possible to load a plugin twice (or more!) on VIP. This patch builds on patches by @batmoo and @davidbinovec to prevent plugins loading twice.

We use a static variable to record all activated plugins each time `wpcom_vip_load_plugin()` is called. This list is checked on every call to ensure the plugin hasn't already been loaded. If it has we alert the VIP team via IRC so we can proactively help VIPs and we also warn admins on the relevant sites using an admin notice.

props philipjohna8c

Differential Revision: https://code.a8c.com/D5367


git-svn-id: https://vip-svn.wordpress.com/plugins@312882 dd790b09-5c1c-0410-8744-875116bcd8a6