Adds a VIP helper function to move the r-login JS to async
Summary: A lot of VIP clients (and us internally) have been talking about r-login recently, and how it hurts their page speeds due to the fact that the JS is render blocking.  This allows r-login to be moved to async JS while still allowing logged in users to use the mapped domain.  More discussion at https://webperfa8c.wordpress.com/2017/08/09/does-removing-r-login-improve-performance/

Test Plan:
To test--

* Add `wpcom_vip_async_rlogin();` to a site's `functions.php` file
* Clear your cookies for a site's mapped domain
* Refresh and watch as you're logged in

I know that r-login is tied to Highlander comments somehow, but I'm not sure how, so this may need to be looked at more deeply.

Reviewers: #vip_developer_squad, olope

Differential Revision: https://code.a8c.com/D6884



git-svn-id: https://vip-svn.wordpress.com/plugins@313103 dd790b09-5c1c-0410-8744-875116bcd8a6