Do not set `is_search` when not in search
WP_Query, when doing initial query parsing, sets `is_search` to `true` when `s` attribute exists, not only when it's non-empty. The `s` attribute also gets backfilled, so we need to reset it, when appropriate.

This commit adds check on whether `is_search` is set or not, and whther the `s` param is empty before unsetting it.

See:
 * https://core.trac.wordpress.org/browser/tags/4.9/src/wp-includes/class-wp-query.php#L773
 * https://core.trac.wordpress.org/browser/tags/4.9/src/wp-includes/class-wp-query.php#L554


git-svn-id: https://vip-svn.wordpress.com/plugins@324031 dd790b09-5c1c-0410-8744-875116bcd8a6