Syndication re-schedule event if current syndication process is running.
Summary: We have reports where even though posts are marked for syndication and published the syndicated posts are never created. In addition testing has shown that the meta of site states is never being set so the error is that syndication is not triggering vs sites not successfully creating the posts. This patch re-schedules the syndication event when triggered via cron.

Test Plan:
1) via shell run `set_transient( 'syn_syndicate_lock', 'loced', 60*5 );
2) Publish a post that is set to syndicate
3) check cron events and review if the syndication event is scheduled for ~ 5 minutes in the future
4) After 5 minutes check if cron event has run, verify posts are syndicated to remote sites

1) via shell run `delete_transient( 'syn_syndicate_lock' );`
2) Publish a post that is set to syndicate
3) verify posts are syndicated to remote sites

Reviewers: #vip_developer_squad, anigel

Reviewed By: #vip_developer_squad, anigel

Subscribers: anigel

Differential Revision: https://code.a8c.com/D10127


git-svn-id: https://vip-svn.wordpress.com/plugins@330882 dd790b09-5c1c-0410-8744-875116bcd8a6